adds nodemon, refactors code into modules
